Changed path handling for results in flixStructure.py initialization to be more reliable
The default value in the 'pathForSaving' parameter in the '__init__' method has been updated. The default path is now set to 'results' instead of '/results'. Furthermore, a line of code that was deleting "/" at the beginning of a path string has been removed. Now a path starting with / is a ABsolute path and one starting with someitng else is a realtive path. Reliable chaccking for C:/ and stuff like this is done by a library.