Set default value of `exists` parameter to 1 and add type checking
In different `__init__` methods, the `exists` parameter default value has been set to 1. Type checking for the `exists` attribute has been added - if it's an integer or a list/array with 0 and 1 only, it will initialize successfully, otherwise it will raise a ValueError. Adjusted the copying of 'exists' information to in-flows and out-flows.